+++ modules/apps/frontend-taglib/frontend-taglib-clay/src/main/java/com/liferay/frontend/taglib/clay/data/set/constants/ClayDataSetConstants.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.frontend.taglib.clay.data.set;
+package com.liferay.frontend.taglib.clay.data.set.constants;
 
 /**
  * @author Marco Leo
